backend/requirement.txt
fastapi 
uvicorn 
python-dotenv 
moorcheh-sdk
pypdf 
python-multipart
langchain-text-splitters
sentence-transformers 
scikit-learn   